requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
autopep8==1.5.7
babel==2.17.0
bark==0.1.5
beartype==0.20.2
beautifulsoup4==4.13.3
biopython==1.85
bleach==6.2.0
blinker==1.9.0
blis==1.3.0
boto3==1.38.13
botocore==1.38.13
Brotli==1.1.0
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
cattrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
comm==0.2.2
confection==0.1.5
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
cymem==2.0.11
daphne==4.2.0
dataclasses-json==0.6.7
debugpy==1.8.13
decorator==4.4.2
defusedxml==0.7.1
discord.py==2.0.0
distlib==0.3.9
distro==1.9.0
Django==5.2.3
django-cors-headers==3.8.0
django-template-lsp==1.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.12.4
dnspython==2.7.0
docopt==0.6.2
docx==0.2.4
einops==0.8.1
email_validator==2.2.0
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastai==2.8.1
fastapi==0.115.12
fastapi-cli==0.0.7
fastcore==1.8.1
fastdownload==0.0.7
fastjsonschema==2.21.1
fastprogress==1.0.3
fasttransform==0.0.2
ffmpy==0.5.0
filelock==3.18.0
filetype==1.2.0
firebase-admin==6.7.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Dance==7.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.2
funcy==2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-speech==2.26.0
google-cloud-storage==2.17.0
google-crc32c==1.7.0
google-genai==1.14.0
google-generativeai==0.8.5
google-pasta==0.2.0
google-resumable-media==2.7.2
google-search==1.1.1
google-search-results==2.4.2
googleapis-common-protos==1.69.2
gradio==5.23.3
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
groq==0.22.0
grpcio==1.71.0
grpcio-status==1.63.0rc1
gTTS==2.5.4
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
incremental==24.7.2
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jellyfish==0.11.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.9.0
kiwisolver==1.4.8
langchain==0.3.22
langchain-community==0.3.20
langchain-core==0.3.54
langchain-google-genai==2.1.3
langchain-text-splitters==0.3.7
langcodes==3.5.0
langsmith==0.3.24
language_data==1.3.0
lazy_loader==0.4
lib50==3.0.12
libclang==18.1.1
llvmlite==0.44.0
lsprotocol==2023.0.1
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
ml_dtypes==0.5.1
mne==1.9.0
more-itertools==10.7.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.3.2
murmurhash==1.0.12
mutagen==1.47.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.31.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
notebook==7.3.3
notebook_shim==0.2.4
num2words==0.5.13
numba==0.61.2
numpy==2.0.2
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.6
plum-dispatch==2.5.7
pooch==1.8.2
preshed==3.0.9
proglog==0.1.10
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.6
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.7.0
pycparser==2.22
pycryptodomex==3.22.0
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
pygls==1.3.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.5
PyNaCl==1.5.0
pyngrok==6.0.0
pyOpenSSL==25.1.0
pyotp==2.9.0
pyparsing==3.2.3
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.1.2
python-dotenv==1.0.0
python-json-logger==3.3.0
python-multipart==0.0.20
pyttsx3==2.98
pytube==15.0.0
pytz==2021.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.14.1
rpds-py==0.23.1
rsa==4.9
ruff==0.11.4
s3transfer==0.12.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.0.2
sentencepiece==0.2.0
serpapi==0.1.4
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
sqlparse==0.4.2
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.1
streamlit==1.36.0
submit50==3.2.0
sympy==1.14.0
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==1.1.0
terminado==0.18.1
tf_keras==2.19.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.3
triton==3.3.0
Twisted==25.5.0
txaio==23.1.1
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
URLObject==2.4.3
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.3
wasabi==1.1.3
watchdog==4.0.2
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
whitenoise==6.9.0
widgetsnbextension==4.0.13
wrapt==1.17.2
xgboost==3.0.2
yarl==1.18.3
yt-dlp==2025.3.27
zope.interface==7.2
zstandard==0.23.0
